+++ GameProject/Assets/Scenes/WoodyPlayerControllTest.unity
@@ -347,6 +347,11 @@ Prefab:
       propertyPath: m_RootOrder
       value: 4
       objectReference: {fileID: 0}
+    - target: {fileID: 114514872716681302, guid: 19b6820faeefb24418bc34b6416d53e9,
+        type: 2}
+      propertyPath: m_Enabled
+      value: 0
+      objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_ParentPrefab: {fileID: 100100000, guid: 19b6820faeefb24418bc34b6416d53e9, type: 2}
   m_IsPrefabParent: 0
