--- GameProject/Assets/Scripts/Player/PlayerAnimation.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ea2b3caf712a9874690b99b237c178b8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
